backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.9
python-multipart==0.0.6
pillow==12.1.1
opencv-python==4.8.1.78
numpy==1.26.4
torch==2.11.0
torchvision==0.26.0
yolov5==7.0.13
pydantic==2.5.0
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
bcrypt==4.1.1
pyjwt==2.9.0
httpx==0.25.0
